Fix sum value type.

+++ sqaodc/cpu/Dot_SIMD.cpp
@@ -220,7 +220,7 @@ double dot_naive(const double *v0, const double *v1, sq::SizeType N) {
 }
 
 float dot_naive(const float *v0, const float *v1, sq::SizeType N) {
-    double sum = 0.;
+    float sum = 0.;
     for (sq::IdxType idx = 0; idx < N; ++idx)
         sum += v0[idx] * v1[idx];
     return sum;
